Give bar control over input generating command.
When the bar is getting input from a loop with a pause, there
can be a noticable delay to visually updating the bar on click
events. By letting bar call an input command it can now update
instantly on a click event and still have a delay so it's not
an idle hog.

Something like:
while true; do date; sleep 10; done | bar -d

Now becomes:
bar -d -c "date" -t 10

Either option may be ommited and the bar will still perform as
before. If a command (-c) is given but not a timeout (-t) then
the timeout defaults to 5 seconds.

Changed compiler standard to gnu99 for use with popen()/pclose()